+++ tests/unit/s2n_x509_validator_test.c
@@ -16,11 +16,6 @@
 #include "s2n_test.h"
 #include "testlib/s2n_testlib.h"
 
-#define CRL_TEST_CHAIN_LEN 2
-
-DEFINE_POINTER_CLEANUP_FUNC(struct s2n_x509_cert*, s2n_x509_cert_free);
-DEFINE_POINTER_CLEANUP_FUNC(struct s2n_x509_crl*, s2n_x509_crl_free);
-
 static int fetch_expired_after_ocsp_timestamp(void *data, uint64_t *timestamp) {
     *timestamp = 7283958536000000000;
     return 0;
@@ -1638,177 +1633,6 @@ int main(int argc, char **argv) {
             EXPECT_TRUE(s2n_x509_trust_store_has_certs(&cfg->trust_store));
             free(cert_chain);
             s2n_config_free(cfg);
-        }
-    }
-
-    /* Test CRL validation */
-    {
-        DEFER_CLEANUP(char *root_crl_pem = malloc(S2N_MAX_TEST_PEM_SIZE), free_char_array_pointer);
-        EXPECT_NOT_NULL(root_crl_pem);
-        EXPECT_SUCCESS(s2n_read_test_pem(S2N_CRL_ROOT_CRL, root_crl_pem, S2N_MAX_TEST_PEM_SIZE));
-        DEFER_CLEANUP(struct s2n_x509_crl *root_crl = NULL, s2n_x509_crl_free_pointer);
-        EXPECT_SUCCESS(s2n_x509_crl_from_pem(root_crl_pem, &root_crl));
-
-        DEFER_CLEANUP(char *intermediate_crl_pem = malloc(S2N_MAX_TEST_PEM_SIZE), free_char_array_pointer);
-        EXPECT_NOT_NULL(intermediate_crl_pem);
-        EXPECT_SUCCESS(s2n_read_test_pem(S2N_CRL_INTERMEDIATE_CRL, intermediate_crl_pem, S2N_MAX_TEST_PEM_SIZE));
-        DEFER_CLEANUP(struct s2n_x509_crl *intermediate_crl = NULL, s2n_x509_crl_free_pointer);
-        EXPECT_SUCCESS(s2n_x509_crl_from_pem(intermediate_crl_pem, &intermediate_crl));
-
-        DEFER_CLEANUP(char *intermediate_revoked_crl_pem = malloc(S2N_MAX_TEST_PEM_SIZE), free_char_array_pointer);
-        EXPECT_NOT_NULL(intermediate_revoked_crl_pem);
-        EXPECT_SUCCESS(s2n_read_test_pem(S2N_CRL_INTERMEDIATE_REVOKED_CRL, intermediate_revoked_crl_pem, S2N_MAX_TEST_PEM_SIZE));
-        DEFER_CLEANUP(struct s2n_x509_crl *intermediate_revoked_crl = NULL, s2n_x509_crl_free_pointer);
-        EXPECT_SUCCESS(s2n_x509_crl_from_pem(intermediate_revoked_crl_pem, &intermediate_revoked_crl));
-
-        /* Initialize a validator with X509s to use for s2n_x509_cert tests */
-        DEFER_CLEANUP(struct s2n_x509_validator retrieved_certs_validator, s2n_x509_validator_wipe);
-        {
-            DEFER_CLEANUP(struct s2n_x509_trust_store trust_store = { 0 }, s2n_x509_trust_store_wipe);
-            s2n_x509_trust_store_init_empty(&trust_store);
-
-            DEFER_CLEANUP(char *root_cert = malloc(S2N_MAX_TEST_PEM_SIZE), free_char_array_pointer);
-            EXPECT_NOT_NULL(root_cert);
-            EXPECT_SUCCESS(s2n_read_test_pem(S2N_CRL_ROOT_CERT, root_cert, S2N_MAX_TEST_PEM_SIZE));
-            EXPECT_SUCCESS(s2n_x509_trust_store_add_pem(&trust_store, root_cert));
-
-            EXPECT_SUCCESS(s2n_x509_validator_init(&retrieved_certs_validator, &trust_store, 0));
-
-            DEFER_CLEANUP(struct s2n_config *config = s2n_config_new(), s2n_config_ptr_free);
-            EXPECT_NOT_NULL(config);
-
-            DEFER_CLEANUP(struct s2n_array *crls = s2n_array_new(sizeof(struct s2n_x509_crl)), s2n_array_free_p);
-            EXPECT_NOT_NULL(crls);
-
-            DEFER_CLEANUP(struct s2n_connection *connection = s2n_connection_new(S2N_CLIENT), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(connection);
-            EXPECT_SUCCESS(s2n_connection_set_config(connection, config));
-            EXPECT_SUCCESS(s2n_set_server_name(connection, "localhost"));
-
-            uint8_t cert_chain_pem[S2N_MAX_TEST_PEM_SIZE];
-            EXPECT_SUCCESS(s2n_read_test_pem(S2N_CRL_NONE_REVOKED_CERT_CHAIN, (char *) cert_chain_pem, S2N_MAX_TEST_PEM_SIZE));
-            DEFER_CLEANUP(struct s2n_stuffer chain_stuffer = { 0 }, s2n_stuffer_free);
-            uint32_t chain_len = write_pem_file_to_stuffer_as_chain(&chain_stuffer, (const char *) cert_chain_pem, S2N_TLS12);
-            EXPECT_TRUE(chain_len > 0);
-            uint8_t *chain_data = s2n_stuffer_raw_read(&chain_stuffer, ( uint32_t )chain_len);
-
-            DEFER_CLEANUP(struct s2n_pkey public_key_out = { 0 }, s2n_pkey_free);
-            EXPECT_SUCCESS(s2n_pkey_zero_init(&public_key_out));
-            s2n_pkey_type pkey_type = S2N_PKEY_TYPE_UNKNOWN;
-            EXPECT_OK(s2n_x509_validator_validate_cert_chain(&retrieved_certs_validator, connection, chain_data,
-                                                             chain_len, &pkey_type, &public_key_out));
-            EXPECT_EQUAL(S2N_PKEY_TYPE_RSA, pkey_type);
-        }
-
-        /* Initializing and freeing empty s2n_x509_cert succeeds */
-        {
-            DEFER_CLEANUP(struct s2n_x509_cert *cert = s2n_x509_cert_new(), s2n_x509_cert_free_pointer);
-            EXPECT_NOT_NULL(cert);
-            EXPECT_NULL(cert->cert);
-        }
-
-        /* Freeing a populated s2n_x509_cert succeeds and doesn't free the internal X509 */
-        {
-            struct s2n_x509_cert *cert = s2n_x509_cert_new();
-            EXPECT_NOT_NULL(cert);
-            EXPECT_NULL(cert->cert);
-
-            EXPECT_TRUE(sk_X509_num(retrieved_certs_validator.cert_chain_from_wire) > 0);
-            X509 *internal_cert = sk_X509_value(retrieved_certs_validator.cert_chain_from_wire, 0);
-            EXPECT_NOT_NULL(internal_cert);
-
-            cert->cert = internal_cert;
-
-            EXPECT_SUCCESS(s2n_x509_cert_free(cert));
-
-            /* Make sure an OpenSSL operation succeeds on the internal X509 */
-            X509_NAME *crl_name = X509_get_issuer_name(internal_cert);
-            POSIX_ENSURE_REF(crl_name);
-        }
-
-        /* Ensure s2n_x509_crl_from_pem produces a valid X509_CRL internally */
-        {
-            EXPECT_NOT_NULL(root_crl->crl);
-
-            /* Make sure an OpenSSL operation succeeds on the internal X509_CRL */
-            X509_NAME *crl_name = X509_CRL_get_issuer(root_crl->crl);
-            POSIX_ENSURE_REF(crl_name);
-        }
-
-        /* s2n_x509_crl_from_pem fails if provided a bad pem */
-        {
-            DEFER_CLEANUP(char *invalid_crl_pem = malloc(S2N_MAX_TEST_PEM_SIZE), free_char_array_pointer);
-            EXPECT_NOT_NULL(invalid_crl_pem);
-            EXPECT_SUCCESS(s2n_read_test_pem(S2N_CRL_ROOT_CRL, invalid_crl_pem, S2N_MAX_TEST_PEM_SIZE));
-
-            /* Change a random byte in the pem to make it invalid */
-            invalid_crl_pem[50] = 1;
-
-            DEFER_CLEANUP(struct s2n_x509_crl *invalid_crl = NULL, s2n_x509_crl_free_pointer);
-            EXPECT_FAILURE_WITH_ERRNO(s2n_x509_crl_from_pem(invalid_crl_pem, &invalid_crl),
-                    S2N_ERR_INTERNAL_LIBCRYPTO_ERROR);
-        }
-
-        /* Certificate issuer hash is retrieved successfully */
-        {
-            DEFER_CLEANUP(struct s2n_x509_cert *cert = s2n_x509_cert_new(), s2n_x509_cert_free_pointer);
-            EXPECT_NOT_NULL(cert);
-
-            EXPECT_TRUE(sk_X509_num(retrieved_certs_validator.cert_chain_from_wire) > 0);
-            X509 *internal_cert = sk_X509_value(retrieved_certs_validator.cert_chain_from_wire, 0);
-            EXPECT_NOT_NULL(internal_cert);
-            cert->cert = internal_cert;
-
-            unsigned long hash = 0;
-            EXPECT_SUCCESS(s2n_x509_cert_get_issuer_hash(cert, &hash));
-            EXPECT_TRUE(hash != 0);
-        }
-
-        /* CRL issuer hash is retrieved successfully */
-        {
-            unsigned long hash = 0;
-            EXPECT_SUCCESS(s2n_x509_crl_get_issuer_hash(root_crl, &hash));
-            EXPECT_TRUE(hash != 0);
-        }
-
-        /* Retrieved hash values for certificates match CRL hashes */
-        {
-            /* Create leaf cert */
-            DEFER_CLEANUP(struct s2n_x509_cert *leaf_cert = s2n_x509_cert_new(), s2n_x509_cert_free_pointer);
-            EXPECT_NOT_NULL(leaf_cert);
-            EXPECT_TRUE(sk_X509_num(retrieved_certs_validator.cert_chain_from_wire) > 1);
-            X509 *internal_leaf_cert = sk_X509_value(retrieved_certs_validator.cert_chain_from_wire, 0);
-            EXPECT_NOT_NULL(internal_leaf_cert);
-            leaf_cert->cert = internal_leaf_cert;
-
-            /* Create intermediate cert */
-            DEFER_CLEANUP(struct s2n_x509_cert *intermediate_cert = s2n_x509_cert_new(), s2n_x509_cert_free_pointer);
-            EXPECT_NOT_NULL(intermediate_cert);
-            X509 *internal_intermediate_cert = sk_X509_value(retrieved_certs_validator.cert_chain_from_wire, 1);
-            EXPECT_NOT_NULL(internal_intermediate_cert);
-            intermediate_cert->cert = internal_intermediate_cert;
-
-            unsigned long intermediate_crl_hash = 0;
-            EXPECT_SUCCESS(s2n_x509_crl_get_issuer_hash(intermediate_crl, &intermediate_crl_hash));
-            EXPECT_TRUE(intermediate_crl_hash != 0);
-
-            unsigned long leaf_cert_hash = 0;
-            EXPECT_SUCCESS(s2n_x509_cert_get_issuer_hash(leaf_cert, &leaf_cert_hash));
-            EXPECT_TRUE(leaf_cert_hash != 0);
-
-            EXPECT_TRUE(intermediate_crl_hash == leaf_cert_hash);
-
-            unsigned long root_crl_hash = 0;
-            EXPECT_SUCCESS(s2n_x509_crl_get_issuer_hash(root_crl, &root_crl_hash));
-            EXPECT_TRUE(root_crl_hash != 0);
-
-            unsigned long intermediate_cert_hash = 0;
-            EXPECT_SUCCESS(s2n_x509_cert_get_issuer_hash(intermediate_cert, &intermediate_cert_hash));
-            EXPECT_TRUE(intermediate_cert_hash != 0);
-
-            EXPECT_TRUE(root_crl_hash == intermediate_cert_hash);
-
-            EXPECT_TRUE(root_crl_hash != leaf_cert_hash);
         }
     }
