+++ tls/s2n_x509_validator.c
@@ -138,100 +138,6 @@ void s2n_x509_trust_store_wipe(struct s2n_x509_trust_store *store) {
     }
 }
 
-struct s2n_x509_cert* s2n_x509_cert_new(void) {
-    DEFER_CLEANUP(struct s2n_blob mem = { 0 }, s2n_free);
-    PTR_GUARD_POSIX(s2n_alloc(&mem, sizeof(struct s2n_x509_cert)));
-
-    struct s2n_x509_cert *cert = (struct s2n_x509_cert*)(void*) mem.data;
-    cert->cert = NULL;
-
-    ZERO_TO_DISABLE_DEFER_CLEANUP(mem);
-    return cert;
-}
-
-int s2n_x509_cert_free(struct s2n_x509_cert *cert) {
-    if (cert) {
-        POSIX_GUARD(s2n_free_object((uint8_t **) &cert, sizeof(struct s2n_x509_cert)));
-    }
-    return S2N_SUCCESS;
-}
-
-int s2n_x509_cert_get_issuer_hash(struct s2n_x509_cert *cert, unsigned long *hash) {
-    POSIX_ENSURE_REF(cert);
-    POSIX_ENSURE_REF(cert->cert);
-
-    unsigned long temp_hash = X509_issuer_name_hash(cert->cert);
-    POSIX_ENSURE(temp_hash != 0, S2N_ERR_INTERNAL_LIBCRYPTO_ERROR);
-
-    *hash = temp_hash;
-
-    return S2N_SUCCESS;
-}
-
-static struct s2n_x509_crl* s2n_x509_crl_new(void) {
-    DEFER_CLEANUP(struct s2n_blob mem = { 0 }, s2n_free);
-    PTR_GUARD_POSIX(s2n_alloc(&mem, sizeof(struct s2n_x509_crl)));
-
-    struct s2n_x509_crl *crl = (struct s2n_x509_crl*)(void*) mem.data;
-    crl->crl = NULL;
-
-    ZERO_TO_DISABLE_DEFER_CLEANUP(mem);
-    return crl;
-}
-
-int s2n_x509_crl_from_pem(char *pem, struct s2n_x509_crl **crl) {
-    POSIX_ENSURE_REF(crl);
-
-    DEFER_CLEANUP(struct s2n_stuffer pem_in_stuffer = {0}, s2n_stuffer_free);
-    DEFER_CLEANUP(struct s2n_stuffer der_out_stuffer = {0}, s2n_stuffer_free);
-
-    POSIX_GUARD(s2n_stuffer_alloc_ro_from_string(&pem_in_stuffer, pem));
-    POSIX_GUARD(s2n_stuffer_growable_alloc(&der_out_stuffer, 2048));
-
-    DEFER_CLEANUP(struct s2n_blob crl_blob = { 0 }, s2n_free);
-
-    POSIX_GUARD(s2n_stuffer_crl_from_pem(&pem_in_stuffer, &der_out_stuffer));
-    POSIX_GUARD(s2n_alloc(&crl_blob, s2n_stuffer_data_available(&der_out_stuffer)));
-    POSIX_GUARD(s2n_stuffer_read(&der_out_stuffer, &crl_blob));
-
-    *crl = s2n_x509_crl_new();
-    POSIX_ENSURE_REF(*crl);
-
-    const uint8_t *data = crl_blob.data;
-    (*crl)->crl = d2i_X509_CRL(NULL, &data, crl_blob.size);
-    POSIX_ENSURE((*crl)->crl != NULL, S2N_ERR_INTERNAL_LIBCRYPTO_ERROR);
-
-    return S2N_SUCCESS;
-}
-
-int s2n_x509_crl_free(struct s2n_x509_crl *crl) {
-    if (crl == NULL) {
-        return S2N_SUCCESS;
-    }
-
-    if (crl->crl) {
-        X509_CRL_free(crl->crl);
-    }
-    POSIX_GUARD(s2n_free_object((uint8_t **) &crl, sizeof(struct s2n_x509_crl)));
-
-    return S2N_SUCCESS;
-}
-
-int s2n_x509_crl_get_issuer_hash(struct s2n_x509_crl *crl, unsigned long *hash) {
-    POSIX_ENSURE_REF(crl);
-    POSIX_ENSURE_REF(crl->crl);
-
-    X509_NAME *crl_name = X509_CRL_get_issuer(crl->crl);
-    POSIX_ENSURE_REF(crl_name);
-
-    unsigned long temp_hash = X509_NAME_hash(crl_name);
-    POSIX_ENSURE(temp_hash != 0, S2N_ERR_INTERNAL_LIBCRYPTO_ERROR);
-
-    *hash = temp_hash;
-
-    return S2N_SUCCESS;
-}
-
 int s2n_x509_validator_init_no_x509_validation(struct s2n_x509_validator *validator) {
     POSIX_ENSURE_REF(validator);
     validator->trust_store = NULL;
